image_retrieval_/requirements.txt
addict==2.4.0
boto3==1.34.133
fastapi==0.112.1
fiftyone==0.24.1
httpx==0.13.3
googletrans==4.0.0-rc1
openai-clip
huggingface_hub==0.23.3
ipdb==0.13.13
matplotlib==3.7.5
# mysql_connector_repackaged==0.3.1
mysql-connector-python
numpy==1.24.4
onnxruntime==1.18.0
opencv_python==4.10.0.82
opencv_python_headless==4.10.0.82
pandas==2.0.3
paramiko==3.4.0
pgvector==0.3.0
Pillow==10.4.0
psycopg2-binary==2.9.9
pycocotools==2.0.7
pydantic==2.8.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.3.2
seaborn==0.13.2
setuptools==69.5.1
SQLAlchemy==2.0.31
sshtunnel==0.4.0
supervision==0.22.0
termcolor==2.4.0
timm==1.0.8
umap-learn==0.5.5
# torch==2.3.1+cu121
# torchvision==0.18.1

torch --extra-index-url https://download.pytorch.org/whl/cu123
torchvision --extra-index-url https://download.pytorch.org/whl/cu123
torchaudio --extra-index-url https://download.pytorch.org/whl/cu123


tqdm==4.66.4
transformers==4.41.2
typer==0.12.4
uvicorn==0.30.6
yapf==0.40.2

python-multipart
